requirements.txt
requests
Pillow
python-dotenv
spidev
gpiozero
lgpio
RPi.GPIO
